Deleting Image Files not Used

+++ controllers/admin.js
@@ -1,5 +1,7 @@
 const { validationResult } = require('express-validator/check');
+const { Error } = require('mongoose');
 const Product = require('../models/product');
+const fileHelper = require('../util/file');
 
 exports.getAddProduct = (req, res, next) => {
   if (!req.session.isLoggedIn) {
@@ -139,6 +141,7 @@ exports.postEditProduct = (req, res, next) => {
       product.price = updatedPrice;
       product.description = updatedDesc;
       if (image) {
+        fileHelper.deleteFile(product.imageUrl);
         product.imageUrl = image.path; 
       }
       return product.save()
@@ -174,7 +177,15 @@ Product.find({ userId: req.user._id })
 
 exports.postDeleteProduct = (req, res, next) => {
   const prodId = req.body.productId;
-  Product.deleteOne({ _id: prodId, userId: req.user._id })
+  Product.findById(prodId)
+    .then(product => {
+      if (!product) {
+        next(new Error('Product not found.'));
+      }
+      
+      fileHelper.deleteFile(product.imageUrl);
+      return Product.deleteOne({ _id: prodId, userId: req.user._id })
+    })
     .then(() => {
       res.redirect('/admin/products');
     })
